experiments/train_router.py
#!/usr/bin/env python3
"""
Train Router Network

Complete training pipeline for the adaptive router network with support for:
- Multi-objective training
- Active learning
- Curriculum learning
- Reinforcement learning fine-tuning
"""

import argparse
import json
from pathlib import Path
from typing import Dict, Any, Optional

import torch
import yaml
from torch.utils.data import DataLoader

import sys
sys.path.insert(0, str(Path(__file__).parent.parent))

from src.router.router_model import AdapterRouter, HierarchicalRouter, DynamicRouter
from src.router.router_trainer import RouterTrainer
from src.router.active_learning import (
    UncertaintySampler, DiversitySampler, HybridSampler
)
from src.router.reinforcement import RLRouterTrainer
from src.utils.logger import get_logger
from src.config.settings import get_settings

logger = get_logger(__name__)


def load_config(config_path: str) -> Dict[str, Any]:
    """Load training configuration from YAML file."""
    with open(config_path, 'r') as f:
        return yaml.safe_load(f)


def create_router(config: Dict[str, Any]) -> torch.nn.Module:
    """Create router model based on configuration."""
    model_config = config.get('model', {})
    
    router_type = model_config.get('type', 'standard')
    
    if router_type == 'hierarchical':
        router = HierarchicalRouter(
            encoder_name=model_config.get('encoder_name', 'prajjwal1/bert-tiny'),
            num_adapters=model_config.get('num_adapters', 4),
            hidden_dim=model_config.get('hidden_dim', 256),
        )
    elif router_type == 'dynamic':
        router = DynamicRouter(
            encoder_name=model_config.get('encoder_name', 'prajjwal1/bert-tiny'),
            num_adapters=model_config.get('num_adapters', 4),
            hidden_dim=model_config.get('hidden_dim', 256),
        )
    else:
        router = AdapterRouter(
            encoder_name=model_config.get('encoder_name', 'prajjwal1/bert-tiny'),
            num_adapters=model_config.get('num_adapters', 4),
            hidden_dim=model_config.get('hidden_dim', 256),
        )
    
    return router


def create_trainer(
    router: torch.nn.Module,
    config: Dict[str, Any],
    output_dir: Path
) -> RouterTrainer:
    """Create router trainer with configuration."""
    training_config = config.get('training', {})
    
    trainer = RouterTrainer(
        router=router,
        adapter_names=config.get('model', {}).get('adapter_names', []),
        output_dir=output_dir,
        learning_rate=training_config.get('learning_rate', 1e-4),
        weight_decay=training_config.get('weight_decay', 0.01),
        num_epochs=training_config.get('num_epochs', 10),
        use_wandb=training_config.get('use_wandb', True),
    )
    
    return trainer


def run_active_learning(
    trainer: RouterTrainer,
    config: Dict[str, Any],
    train_data: Any,
    unlabeled_data: Any
) -> None:
    """Run active learning loop."""
    al_config = config.get('active_learning', {})
    
    if not al_config.get('enabled', False):
        return
    
    strategy = al_config.get('strategy', 'hybrid')
    
    if strategy == 'uncertainty':
        sampler = UncertaintySampler(
            model=trainer.router,
            method=al_config.get('uncertainty_method', 'entropy'),
        )
    elif strategy == 'diversity':
        sampler = DiversitySampler(
            model=trainer.router,
            method=al_config.get('diversity_method', 'kmeans'),
        )
    else:
        sampler = HybridSampler(
            model=trainer.router,
            uncertainty_weight=al_config.get('uncertainty_weight', 0.5),
        )
    
    query_size = al_config.get('query_size', 100)
    max_iterations = al_config.get('max_iterations', 10)
    
    logger.info(f"Starting active learning with {strategy} strategy")
    
    for iteration in range(max_iterations):
        # Select samples
        indices = sampler.select_samples(unlabeled_data, n_samples=query_size)
        
        # Simulate labeling (in practice, this would involve human annotation)
        new_samples = [unlabeled_data[i] for i in indices]
        
        # Add to training data and retrain
        # train_data.extend(new_samples)
        # trainer.train(train_data)
        
        logger.info(f"Active learning iteration {iteration + 1}: added {len(indices)} samples")


def run_reinforcement_learning(
    router: torch.nn.Module,
    config: Dict[str, Any],
    output_dir: Path
) -> None:
    """Run reinforcement learning fine-tuning."""
    rl_config = config.get('reinforcement', {})
    
    if not rl_config.get('enabled', False):
        return
    
    algorithm = rl_config.get('algorithm', 'ppo')
    
    rl_trainer = RLRouterTrainer(
        router=router,
        algorithm=algorithm,
        learning_rate=rl_config.get('learning_rate', 1e-5),
        gamma=rl_config.get('gamma', 0.99),
        clip_epsilon=rl_config.get('clip_epsilon', 0.2),
    )
    
    logger.info(f"Starting RL fine-tuning with {algorithm}")
    
    # Run RL training
    # rl_trainer.train(environment, num_episodes=rl_config.get('num_episodes', 1000))


def main():
    parser = argparse.ArgumentParser(description='Train Adaptive Router')
    parser.add_argument(
        '--config', 
        type=str, 
        default='configs/router/training_config.yaml',
        help='Path to training configuration'
    )
    parser.add_argument(
        '--output-dir',
        type=str,
        default='outputs/router',
        help='Output directory for checkpoints'
    )
    parser.add_argument(
        '--resume',
        type=str,
        default=None,
        help='Path to checkpoint to resume from'
    )
    parser.add_argument(
        '--active-learning',
        action='store_true',
        help='Enable active learning'
    )
    parser.add_argument(
        '--reinforcement-learning',
        action='store_true',
        help='Enable RL fine-tuning'
    )
    
    args = parser.parse_args()
    
    # Load configuration
    config = load_config(args.config)
    
    # Override with command line args
    if args.active_learning:
        config.setdefault('active_learning', {})['enabled'] = True
    if args.reinforcement_learning:
        config.setdefault('reinforcement', {})['enabled'] = True
    
    output_dir = Path(args.output_dir)
    output_dir.mkdir(parents=True, exist_ok=True)
    
    # Save config
    with open(output_dir / 'config.yaml', 'w') as f:
        yaml.dump(config, f)
    
    logger.info(f"Training router with config: {args.config}")
    logger.info(f"Output directory: {output_dir}")
    
    # Create router
    router = create_router(config)
    logger.info(f"Created router: {type(router).__name__}")
    
    # Create trainer
    trainer = create_trainer(router, config, output_dir)
    
    # Resume if specified
    if args.resume:
        trainer.load_checkpoint(args.resume)
        logger.info(f"Resumed from checkpoint: {args.resume}")
    
    # Train
    logger.info("Starting training...")
    # trainer.train(train_dataloader, eval_dataloader)
    
    # Active learning
    if config.get('active_learning', {}).get('enabled', False):
        logger.info("Running active learning...")
        # run_active_learning(trainer, config, train_data, unlabeled_data)
    
    # RL fine-tuning
    if config.get('reinforcement', {}).get('enabled', False):
        logger.info("Running RL fine-tuning...")
        # run_reinforcement_learning(router, config, output_dir)
    
    # Save final model
    final_path = output_dir / 'final_router'
    torch.save(router.state_dict(), final_path / 'model.pt')
    logger.info(f"Training complete. Model saved to {final_path}")


if __name__ == '__main__':
    main()
