requirements.txt
# Core ML/DL
torch>=2.1.0
transformers>=4.36.0
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.25.0
datasets>=2.15.0

# Training
deepspeed>=0.12.0
flash-attn>=2.3.0
wandb>=0.16.0

# Inference
vllm>=0.2.6
ray>=2.8.0

# API & Serving
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
redis>=5.0.0
celery>=5.3.0

# Data Processing
pandas>=2.1.0
numpy>=1.24.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Evaluation
scikit-learn>=1.3.0
scipy>=1.11.0
nltk>=3.8.0

# Monitoring
prometheus-client>=0.19.0
psutil>=5.9.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
black>=23.11.0
isort>=5.12.0
mypy>=1.7.0
pre-commit>=3.5.0

# AWS
boto3>=1.29.0

# API Clients
openai>=1.0.0
anthropic>=0.8.0
tenacity>=8.2.0

# Configuration
pyyaml>=6.0.0
python-dotenv>=1.0.0
